deployments/nsfw_detector/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130

ray[serve]>=2.9.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
Pillow>=10.0.0
httpx>=0.25.0
dotenv
